TuHA: confirm booking

+++ src/main/java/com/higgsup/base/dto/QuoteResultDTO.java
@@ -2,12 +2,14 @@ package com.higgsup.base.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class QuoteResultDTO {
-    private Double baseCharge;
-    private Double fuelSurcharge;
-    private Double dangerousCharge;
-    private Double totalCharge;
+    private BigDecimal baseCharge;
+    private BigDecimal fuelSurcharge;
+    private BigDecimal dangerousCharge;
+    private BigDecimal totalCharge;
     private Double totalWeight;
     private String weightType;
 }
